vault-sentinel-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
sse-starlette==1.8.2
pandas==2.1.3
scikit-learn==1.3.2
shap==0.43.0
reportlab==4.0.7
google-generativeai==0.3.2
faker==20.1.0
python-multipart==0.0.6
celery==5.3.4
aiosmtplib==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
